Clean up components directory (#14)
