+++ src/components/TitleCard/index.js
@@ -1,0 +1,3 @@
+import TitleCard from './TitleCard';
+
+export default TitleCard;
